+++ src/neo4j/_async_compat/util.py
@@ -18,6 +18,7 @@ from __future__ import annotations
 
 import asyncio
 import inspect
+import itertools
 import traceback
 import typing as t
 from functools import wraps
@@ -27,6 +28,11 @@ if t.TYPE_CHECKING:
     import typing_extensions as te
 
     _T = t.TypeVar("_T")
+    _T1 = t.TypeVar("_T1")
+    _T2 = t.TypeVar("_T2")
+    _T3 = t.TypeVar("_T3")
+    _T4 = t.TypeVar("_T4")
+    _T5 = t.TypeVar("_T5")
     _P = te.ParamSpec("_P")
 
 
@@ -38,16 +44,24 @@ __all__ = [
 
 class AsyncUtil:
     @staticmethod
-    async def iter(it):
+    async def chain_iter(
+        its: t.Iterable[t.AsyncIterable[_T]],
+    ) -> t.AsyncIterable[_T]:
+        for it in its:
+            async for x in it:
+                yield x
+
+    @staticmethod
+    async def iter(it: t.AsyncIterable[_T]) -> t.AsyncIterator[_T]:
         async for x in it:
             yield x
 
     @staticmethod
-    async def next(it):
+    async def next(it: t.AsyncIterator[_T]) -> _T:
         return await it.__anext__()
 
     @staticmethod
-    async def list(it):
+    async def list(it: t.AsyncIterable[_T]) -> list[_T]:
         return [x async for x in it]
 
     @staticmethod
@@ -76,6 +90,12 @@ class AsyncUtil:
         return None
 
     @staticmethod
+    async def maybe_await(value: _T1 | t.Awaitable[_T2]) -> _T1 | _T2:
+        if inspect.isawaitable(value):
+            return await value
+        return t.cast(_T1, value)
+
+    @staticmethod
     def shielded(coro_function):
         assert asyncio.iscoroutinefunction(coro_function)
 
@@ -85,7 +105,7 @@ class AsyncUtil:
 
         return shielded_function
 
-    is_async_code: t.ClassVar = True
+    is_async_code: t.ClassVar[bool] = True
 
     @staticmethod
     def extract_stack(limit=None):
@@ -97,6 +117,7 @@ class AsyncUtil:
 
 
 class Util:
+    chain_iter = itertools.chain.from_iterable
     iter: t.ClassVar = iter
     next: t.ClassVar = next
     list: t.ClassVar = list
@@ -118,10 +139,14 @@ class Util:
         return None
 
     @staticmethod
+    def maybe_await(value: _T) -> _T:
+        return value
+
+    @staticmethod
     def shielded(coro_function):
         return coro_function
 
-    is_async_code: t.ClassVar = False
+    is_async_code: t.ClassVar[bool] = False
 
     @staticmethod
     def extract_stack(limit=None):