+++ src/neo4j/_sync/io/_pool.py
@@ -143,6 +143,12 @@ class IOPool(abc.ABC):
         with self.lock:
             return self._ssr_feature_tracker.has_feature
 
+    @abc.abstractmethod
+    def _pool_key_by_address(self, address): ...
+
+    def _pool_key_by_connection(self, connection):
+        return self._pool_key_by_address(connection.unresolved_address)
+
     def __enter__(self):
         return self
 
@@ -152,9 +158,9 @@ class IOPool(abc.ABC):
     def get_auth(self):
         return Util.callback(self.pool_config.auth.get_auth)
 
-    def _acquire_from_pool(self, address):
+    def _acquire_from_pool(self, pool_key):
         with self.lock:
-            for connection in self.connections[address]:
+            for connection in self.connections[pool_key]:
                 if connection.in_use:
                     continue
                 connection.pool = self
@@ -163,12 +169,12 @@ class IOPool(abc.ABC):
         return None  # no free connection available
 
     def _remove_connection(self, connection):
-        address = connection.unresolved_address
+        pool_key = self._pool_key_by_connection(connection)
         with self.lock:
             log.debug(
                 "[#%04X]  _: <POOL> remove connection from pool %r %s",
                 connection.local_port,
-                address,
+                pool_key,
                 connection.connection_id,
             )
             # If closure fails (e.g. because the server went
@@ -176,27 +182,27 @@ class IOPool(abc.ABC):
             # be removed. Therefore, we silently ignore if the
             # connection isn't in the pool anymore.
             with suppress(ValueError):
-                self.connections.get(address, []).remove(connection)
+                self.connections.get(pool_key, []).remove(connection)
                 self._ssr_feature_tracker.remove_connection(connection)
 
-    def _add_connections(self, address, *connections):
+    def _add_connections(self, pool_key, *connections):
         with self.lock:
-            self.connections[address].extend(connections)
+            self.connections[pool_key].extend(connections)
             for connection in connections:
                 self._ssr_feature_tracker.add_connection(connection)
 
-    def _remove_connections(self, address, *connections):
+    def _remove_connections(self, pool_key, *connections):
         with self.lock:
-            existing_connections = self.connections.get(address, [])
+            existing_connections = self.connections.get(pool_key, [])
             for connection in connections:
                 existing_connections.remove(connection)
                 self._ssr_feature_tracker.remove_connection(connection)
 
     def _acquire_from_pool_checked(
-        self, address, health_check, deadline
+        self, pool_key, health_check, deadline
     ):
         while not deadline.expired():
-            connection = self._acquire_from_pool(address)
+            connection = self._acquire_from_pool(pool_key)
             if not connection:
                 return None  # no free connection available
             if not health_check(connection, deadline):
@@ -222,7 +228,7 @@ class IOPool(abc.ABC):
                 return connection
         return None
 
-    def _acquire_new_later(self, address, auth, deadline):
+    def _acquire_new_later(self, pool_key, address, auth, deadline):
         def connection_creator():
             released_reservation = False
             try:
@@ -249,25 +255,25 @@ class IOPool(abc.ABC):
                 connection.pool = self
                 connection.in_use = True
                 with self.lock:
-                    self.connections_reservations[address] -= 1
+                    self.connections_reservations[pool_key] -= 1
                     released_reservation = True
-                    self._add_connections(address, connection)
+                    self._add_connections(pool_key, connection)
                 return connection
             finally:
                 if not released_reservation:
                     with self.lock:
-                        self.connections_reservations[address] -= 1
+                        self.connections_reservations[pool_key] -= 1
 
         max_pool_size = self.pool_config.max_connection_pool_size
         infinite_pool_size = max_pool_size < 0 or max_pool_size == float("inf")
         with self.lock:
-            connections = self.connections[address]
+            connections = self.connections[pool_key]
             pool_size = (
-                len(connections) + self.connections_reservations[address]
+                len(connections) + self.connections_reservations[pool_key]
             )
             if infinite_pool_size or pool_size < max_pool_size:
                 # there's room for a new connection
-                self.connections_reservations[address] += 1
+                self.connections_reservations[pool_key] += 1
                 return connection_creator
         return None
 
@@ -322,6 +328,7 @@ class IOPool(abc.ABC):
             auth = AcquisitionAuth(None)
         force_auth = auth.force_auth
         auth = auth.auth
+        pool_key = self._pool_key_by_address(address)
         if liveness_check_timeout is None:
             liveness_check_timeout = self.pool_config.liveness_check_timeout
 
@@ -349,7 +356,7 @@ class IOPool(abc.ABC):
         while True:
             # try to find a free connection in the pool
             connection = self._acquire_from_pool_checked(
-                address, health_check, deadline
+                pool_key, health_check, deadline
             )
             if connection:
                 log.debug(
@@ -384,6 +391,7 @@ class IOPool(abc.ABC):
             # all connections in pool are in-use
             with self.lock:
                 connection_creator = self._acquire_new_later(
+                    pool_key,
                     address,
                     auth,
                     deadline,
@@ -507,10 +515,10 @@ class IOPool(abc.ABC):
         if cancelled is not None:
             raise cancelled
 
-    def in_use_connection_count(self, address):
-        """Count the connections currently in use to a given address."""
+    def in_use_connection_count(self, pool_key):
+        """Count the connections currently in use to a given pool_key."""
         with self.lock:
-            connections = self.connections.get(address, ())
+            connections = self.connections.get(pool_key, ())
             return sum(connection.in_use for connection in connections)
 
     def mark_all_stale(self):
@@ -542,9 +550,10 @@ class IOPool(abc.ABC):
 
         If present, close all idle connections to that address.
         """
+        pool_key = self._pool_key_by_address(address)
         with self.lock:
             try:
-                connections = self.connections[address]
+                connections = self.connections[pool_key]
             except KeyError:  # already removed from the connection pool
                 return
             closable_connections = [
@@ -553,9 +562,9 @@ class IOPool(abc.ABC):
             # First remove all connections in question, then try to close them.
             # If closing of a connection fails, we will end up in this method
             # again.
-            self._remove_connections(address, *closable_connections)
-            if not self.connections[address]:
-                del self.connections[address]
+            self._remove_connections(pool_key, *closable_connections)
+            if not self.connections[pool_key]:
+                del self.connections[pool_key]
 
         self._close_connections(closable_connections)
 
@@ -567,14 +576,14 @@ class IOPool(abc.ABC):
     def on_neo4j_error(self, error, connection):
         assert isinstance(error, Neo4jError)
         if error._unauthenticates_all_connections():
-            address = connection.unresolved_address
+            pool_key = self._pool_key_by_connection(connection)
             log.debug(
                 "[#0000]  _: <POOL> mark all connections to %r as "
                 "unauthenticated",
-                address,
+                pool_key,
             )
             with self.lock:
-                for connection_ in self.connections.get(address, ()):
+                for connection_ in self.connections.get(pool_key, ()):
                     connection_.mark_unauthenticated()
         if error._has_security_code():
             handled = Util.callback(
@@ -596,8 +605,8 @@ class IOPool(abc.ABC):
             with self.lock:
                 connections = [
                     connection
-                    for address in list(self.connections)
-                    for connection in self.connections.pop(address, ())
+                    for pool_key in list(self.connections)
+                    for connection in self.connections.pop(pool_key, ())
                 ]
                 for connection in connections:
                     self._ssr_feature_tracker.remove_connection(connection)
@@ -609,12 +618,15 @@ class IOPool(abc.ABC):
 class BoltPool(IOPool):
     is_direct_pool = True
 
+    def _pool_key_by_address(self, address):
+        return None
+
     @classmethod
-    def open(cls, address, *, pool_config, workspace_config):
+    def open(cls, multi_addr, *, pool_config, workspace_config):
         """
         Create a new BoltPool.
 
-        :param address:
+        :param multi_addr:
         :param pool_config:
         :param workspace_config:
         :returns: BoltPool
@@ -629,16 +641,16 @@ class BoltPool(IOPool):
                 pool_config=pool_config,
             )
 
-        pool = cls(opener, pool_config, workspace_config, address)
-        log.debug("[#0000]  _: <POOL> created, direct address %r", address)
+        pool = cls(opener, pool_config, workspace_config, multi_addr)
+        log.debug("[#0000]  _: <POOL> created, direct address %r", multi_addr)
         return pool
 
-    def __init__(self, opener, pool_config, workspace_config, address):
+    def __init__(self, opener, pool_config, workspace_config, multi_addr):
         super().__init__(opener, pool_config, workspace_config)
-        self.address = address
+        self.multi_addr = multi_addr
 
     def __repr__(self):
-        return f"<{self.__class__.__name__} address={self.address!r}>"
+        return f"<{self.__class__.__name__} address={self.multi_addr!r}>"
 
     def acquire(
         self,
@@ -660,8 +672,12 @@ class BoltPool(IOPool):
         )
         deadline = Deadline.from_timeout_or_deadline(timeout)
         return self._acquire(
-            self.address, auth, deadline, liveness_check_timeout
+            self.multi_addr, auth, deadline, liveness_check_timeout
         )
+
+    def deactivate(self, address):
+        log.debug("[#0000]  _: <POOL> deactivating (direct pool)")
+        super().deactivate(address)
 
 
 class Neo4jPool(IOPool):
@@ -669,10 +685,14 @@ class Neo4jPool(IOPool):
 
     is_direct_pool = False
 
+    def _pool_key_by_address(self, address):
+        return address
+
+    def _pool_key_by_connection(self, connection):
+        return connection.unresolved_address
+
     @classmethod
-    def open(
-        cls, *addresses, pool_config, workspace_config, routing_context=None
-    ):
+    def open(cls, multi_addr, pool_config, workspace_config):
         """
         Create a new Neo4jPool.
 
@@ -682,41 +702,38 @@ class Neo4jPool(IOPool):
         :param routing_context:
         :returns: Neo4jPool
         """
-        address = addresses[0]
-        if routing_context is None:
-            routing_context = {}
-        elif "address" in routing_context:
-            raise ConfigurationError(
-                "The key 'address' is reserved for routing context."
-            )
-        routing_context["address"] = str(address)
 
         def opener(addr, auth_manager, deadline):
             return Bolt.open(
                 addr,
                 auth_manager=auth_manager,
                 deadline=deadline,
-                routing_context=routing_context,
+                routing_context={
+                    "address": str(addr._unresolved),
+                    **multi_addr.routing_context,
+                },
                 pool_config=pool_config,
             )
 
-        pool = cls(opener, pool_config, workspace_config, address)
-        log.debug("[#0000]  _: <POOL> created, routing address %r", address)
+        pool = cls(opener, pool_config, workspace_config, multi_addr)
+        log.debug(
+            "[#0000]  _: <POOL> created, routing addresses %r", multi_addr
+        )
         return pool
 
-    def __init__(self, opener, pool_config, workspace_config, address):
+    def __init__(self, opener, pool_config, workspace_config, multi_addr):
         """
         Initialize a Neo4jPool.
 
         :param opener:
         :param pool_config:
         :param workspace_config:
-        :param address:
+        :param multi_addr:
         """
         super().__init__(opener, pool_config, workspace_config)
         # Each database have a routing table, the default database is a special
         # case.
-        self.address = address
+        self.multi_addr = multi_addr
         self.routing_tables = {}
         self.refresh_lock = RLock()
         self.is_direct_pool = False
@@ -728,13 +745,14 @@ class Neo4jPool(IOPool):
         :returns: The representation
         :rtype: str
         """
-        return f"<{self.__class__.__name__} address={self.address!r}>"
+        return f"<{self.__class__.__name__} addresses={self.multi_addr!r}>"
 
     def get_or_create_routing_table(self, database):
         with self.refresh_lock:
             if database not in self.routing_tables:
+                addresses = self.multi_addr.addresses()
                 self.routing_tables[database] = RoutingTable(
-                    database=database, routers=[self.address]
+                    database=database, routers=addresses
                 )
             return self.routing_tables[database]
 
@@ -887,14 +905,16 @@ class Neo4jPool(IOPool):
         :returns: True if the routing table is successfully updated,
         otherwise False
         """
-        if routers:
-            log.debug(
-                (
-                    "[#0000]  _: <POOL> "
-                    "attempting to update routing table from %s"
-                ),
-                ", ".join(map(repr, routers)),
-            )
+        if not routers:
+            return False
+
+        log.debug(
+            (
+                "[#0000]  _: <POOL> "
+                "attempting to update routing table from %s"
+            ),
+            ", ".join(map(repr, routers)),
+        )
         for router in routers:
             for address in NetworkUtil.resolve_address(
                 router, resolver=self.pool_config.resolver
@@ -952,24 +972,25 @@ class Neo4jPool(IOPool):
 
         :raise neo4j.exceptions.ServiceUnavailable:
         """
+        addresses = self.multi_addr.addresses()
         with self.refresh_lock:
             routing_table = self.get_routing_table(database)
             if routing_table is not None:
                 # copied because it can be modified
                 existing_routers = set(routing_table.routers)
-                prefer_initial_routing_address = (
+                prefer_initial_routing_addresses = (
                     routing_table.initialized_without_writers
                 )
             else:
-                existing_routers = {self.address}
-                prefer_initial_routing_address = True
+                existing_routers = set()
+                prefer_initial_routing_addresses = True
 
             if (
-                prefer_initial_routing_address
+                prefer_initial_routing_addresses
                 and
                 # TODO: Test this state
                 self._update_routing_table_from(
-                    self.address,
+                    *addresses,
                     database=database,
                     imp_user=imp_user,
                     bookmarks=bookmarks,
@@ -981,7 +1002,7 @@ class Neo4jPool(IOPool):
                 # Why is only the first initial routing address used?
                 return
             if self._update_routing_table_from(
-                *(existing_routers - {self.address}),
+                *(existing_routers - set(addresses)),
                 database=database,
                 imp_user=imp_user,
                 bookmarks=bookmarks,
@@ -992,9 +1013,9 @@ class Neo4jPool(IOPool):
                 return
 
             if (
-                not prefer_initial_routing_address
+                not prefer_initial_routing_addresses
                 and self._update_routing_table_from(
-                    self.address,
+                    *addresses,
                     database=database,
                     imp_user=imp_user,
                     bookmarks=bookmarks,
@@ -1016,7 +1037,7 @@ class Neo4jPool(IOPool):
 
         servers = set.union(
             *(rt.servers() for rt in routing_tables),
-            self.address,
+            self.multi_addr.addresses(),
         )
         for address in list(self.connections):
             if address._unresolved not in servers: