+++ tests/unit/async_/io/test_neo4j_pool.py
@@ -22,6 +22,7 @@ from neo4j import (
     READ_ACCESS,
     WRITE_ACCESS,
 )
+from neo4j._async.addressing import _AsyncMultiAddress
 from neo4j._async.config import AsyncPoolConfig
 from neo4j._async.io import (
     AcquisitionDatabase,
@@ -142,9 +143,12 @@ def _auth_manager(auth):
     return AsyncAuthManagers.static(auth)
 
 
+POOL_ADDRESS = _AsyncMultiAddress("neo4j", (ROUTER1_ADDRESS,))
+
+
 def _simple_pool(opener) -> AsyncNeo4jPool:
     return AsyncNeo4jPool(
-        opener, _pool_config(), WorkspaceConfig(), ROUTER1_ADDRESS
+        opener, _pool_config(), WorkspaceConfig(), POOL_ADDRESS
     )
 
 
@@ -562,9 +566,11 @@ async def test_failing_opener_leaves_connections_in_use_alone(opener):
 async def test__acquire_new_later_with_room(opener):
     config = _pool_config()
     config.max_connection_pool_size = 1
-    pool = AsyncNeo4jPool(opener, config, WorkspaceConfig(), ROUTER1_ADDRESS)
+    pool = AsyncNeo4jPool(opener, config, WorkspaceConfig(), POOL_ADDRESS)
     assert pool.connections_reservations[READER1_ADDRESS] == 0
-    creator = pool._acquire_new_later(READER1_ADDRESS, None, Deadline(1))
+    creator = pool._acquire_new_later(
+        READER1_ADDRESS, READER1_ADDRESS, None, Deadline(1)
+    )
     assert pool.connections_reservations[READER1_ADDRESS] == 1
     assert callable(creator)
     if AsyncUtil.is_async_code:
@@ -575,11 +581,13 @@ async def test__acquire_new_later_with_room(opener):
 async def test__acquire_new_later_without_room(opener):
     config = _pool_config()
     config.max_connection_pool_size = 1
-    pool = AsyncNeo4jPool(opener, config, WorkspaceConfig(), ROUTER1_ADDRESS)
+    pool = AsyncNeo4jPool(opener, config, WorkspaceConfig(), POOL_ADDRESS)
     _ = await pool.acquire(READ_ACCESS, 30, TEST_DB1, None, None, None)
     # pool is full now
     assert pool.connections_reservations[READER1_ADDRESS] == 0
-    creator = pool._acquire_new_later(READER1_ADDRESS, None, Deadline(1))
+    creator = pool._acquire_new_later(
+        READER1_ADDRESS, READER1_ADDRESS, None, Deadline(1)
+    )
     assert pool.connections_reservations[READER1_ADDRESS] == 0
     assert creator is None
 
@@ -591,11 +599,13 @@ async def test_passes_pool_config_to_connection(mocker):
     pool_config = AsyncPoolConfig()
     workspace_config = WorkspaceConfig()
     pool = AsyncNeo4jPool.open(
-        mocker.Mock, pool_config=pool_config, workspace_config=workspace_config
+        mocker.MagicMock(),
+        pool_config=pool_config,
+        workspace_config=workspace_config,
     )
 
     _ = await pool._acquire(
-        mocker.Mock, None, Deadline.from_timeout_or_deadline(30), None
+        mocker.Mock(), None, Deadline.from_timeout_or_deadline(30), None
     )
 
     bolt_mock.assert_awaited_once()
@@ -628,7 +638,10 @@ async def test_discovery_is_retried(custom_routing_opener, error):
         opener,
         _pool_config(),
         WorkspaceConfig(),
-        ResolvedAddress(("1.2.3.1", 9999), host_name="host"),
+        _AsyncMultiAddress(
+            "neo4j",
+            (ResolvedAddress(("1.2.3.1", 9999), host_name="host"),),
+        ),
     )
     cx1 = await pool.acquire(READ_ACCESS, 30, TEST_DB1, None, None, None)
     await pool.release(cx1)
@@ -680,7 +693,10 @@ async def test_fast_failing_discovery(custom_routing_opener, error):
         opener,
         _pool_config(),
         WorkspaceConfig(),
-        ResolvedAddress(("1.2.3.1", 9999), host_name="host"),
+        _AsyncMultiAddress(
+            "neo4j",
+            (ResolvedAddress(("1.2.3.1", 9999), host_name="host"),),
+        ),
     )
     cx1 = await pool.acquire(READ_ACCESS, 30, TEST_DB1, None, None, None)
     await pool.release(cx1)
@@ -726,7 +742,7 @@ async def test_connection_error_callback(
         auth_manager, "handle_security_exception", autospec=True
     )
     config.auth = auth_manager
-    pool = AsyncNeo4jPool(opener, config, WorkspaceConfig(), ROUTER1_ADDRESS)
+    pool = AsyncNeo4jPool(opener, config, WorkspaceConfig(), POOL_ADDRESS)
     cxs_read = [
         await pool.acquire(READ_ACCESS, 30, TEST_DB1, None, None, None)
         for _ in range(5)
@@ -769,7 +785,7 @@ async def test_pool_closes_connections_dropped_from_rt(custom_routing_opener):
     opener = custom_routing_opener(get_readers=get_readers)
 
     pool = AsyncNeo4jPool(
-        opener, _pool_config(), WorkspaceConfig(), ROUTER1_ADDRESS
+        opener, _pool_config(), WorkspaceConfig(), POOL_ADDRESS
     )
     cx1 = await pool.acquire(READ_ACCESS, 30, TEST_DB1, None, None, None)
     assert cx1.unresolved_address == READER1_ADDRESS
@@ -807,7 +823,7 @@ async def test_pool_does_not_close_connections_dropped_from_rt_for_other_server(
     opener = custom_routing_opener(get_readers=get_readers)
 
     pool = AsyncNeo4jPool(
-        opener, _pool_config(), WorkspaceConfig(), ROUTER1_ADDRESS
+        opener, _pool_config(), WorkspaceConfig(), POOL_ADDRESS
     )
     cx1 = await pool.acquire(READ_ACCESS, 30, TEST_DB1, None, None, None)
     await pool.release(cx1)
@@ -858,7 +874,7 @@ async def test_tracks_ssr_connection_hints(custom_routing_opener):
 
     opener = custom_routing_opener(on_open=on_open)
     pool = AsyncNeo4jPool(
-        opener, _pool_config(), WorkspaceConfig(), ROUTER1_ADDRESS
+        opener, _pool_config(), WorkspaceConfig(), POOL_ADDRESS
     )
 
     # no connection in pool => cannot know => defensive assumption: off